Refactor: 컴포넌트명 변경 (PointItem, FriendManageItem)
