+++ app/pointHistory/page.tsx
@@ -1,4 +1,4 @@
-import PointItem from "@/components/PointItem";
+import PointHistoryList from "@/components/PointHistoryList";
 import { createDummyData } from "@/public/dummy";
 
 // TODO: dummy data 추후 실제 데이터로 변경
@@ -8,7 +8,7 @@ export default function PointHistoryPage() {
   return (
     <div>
       {dummyData.map((item) => (
-        <PointItem key={`${item.date}_${item.amount}`} item={item} />
+        <PointHistoryList key={`${item.date}_${item.amount}`} item={item} />
       ))}
     </div>
   );
