+++ components/PointHistoryList.tsx
